% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas-funcoes.R
\name{def_pol}
\alias{def_pol}
\title{Define se os pontos pertencem a um polígono}
\usage{
def_pol(x, y, pol)
}
\arguments{
\item{x}{Vetor com a coordenada x do ponto}

\item{y}{Vetor com a coordenada y do ponto}

\item{pol}{matriz do polígono}
}
\value{
o retorno da função será um vetor lógico dizendo que os pontos
pertencem ou não ao polígono especificado por pol
}
\description{
Verifica para um ou mais pontos se eles se enquadram em um determinado polígono.
}
\details{
Função baseada na função point.in.pol do pacote sp, utilizada
para filtrar os pontos pertencentes aos polígonos dos estados,
regiões e muicípios do Brasil. Os polígonos são oriundos do pacote geobr.
}
