% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{oco2}
\alias{oco2}
\title{Concentração de CO2 atmosférico}
\format{
Uma lista que descreve as colunas: esta base tem X linahs e X colunas
\describe{
\item{col1}{Descrição}
\item{col1}{Descrição}

}
}
\source{
\url{www.fcav.unesp.br}
}
\usage{
oco2
}
\description{
Descrição da base
}
\keyword{datasets}
