% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{data_fco2}
\alias{data_fco2}
\title{Dados de emissão de CO2 do solo e atributos relacionados}
\format{
Uma tibble, contendo 15397 linhas e 39 colunas
\describe{
\item{experimento}{Nos experimentos classificados como \emph{Espacial} uma malha
amostral foi instalada em campo como o objetivo de avaliar a variabilidade espacial
dos atributos do solo por meio de técnicas geoestatísticas; nos experimentos classificados
como \strong{Temporais}, a respiração do solo foi avaliada em diferentes tratamentos (usos do solo,
manejos ou preparos)}
\item{data}{descrição}
\item{preparo_solo}{descrição}
\item{tratamento}{descrição}
\item{revolvimento_solo}{descrição}
\item{data_preparo}{descrição}
\item{conversao}{descrição}
\item{cobertura}{descrição}
\item{cultura}{descrição}
\item{x}{descrição}
\item{y}{descrição}
\item{longitude_muni}{descrição}
\item{latitude_muni}{descrição}
\item{estado}{descrição}
\item{municipio}{descrição}
\item{ID}{descrição}
\item{prof}{descrição}
\item{FCO2}{\eqn{\mu mol \; CO_2 m^{-2}}
}
\item{Ts}{descrição}
\item{Us}{descrição}
\item{pH}{descrição}
\item{MO}{descrição}
\item{P}{descrição}
\item{K}{descrição}
\item{Ca}{descrição}
\item{Mg}{descrição}
\item{H_Al}{descrição}
\item{SB}{descrição}
\item{CTC}{descrição}
\item{V}{descrição}
\item{Ds}{descrição}
\item{Macro}{descrição}
\item{Micro}{descrição}
\item{VTP}{descrição}
\item{PLA}{descrição}
\item{AT}{descrição}
\item{SILTE}{descrição}
\item{ARG}{descrição}
\item{HLIFS}{descrição}
}
}
\source{
\url{www.fcav.unesp.br}
}
\usage{
data_fco2
}
\description{
Base de dados contendo valores de emissão de CO2 e atributos físicos e químicos do solo para diferentes experimentos conduzidos no Brasil
}
\keyword{datasets}
